+++ screens/AllUserDog/DogsScreen.jsx
@@ -1,10 +1,8 @@
 import React, {useState} from 'react';
 import { StyleSheet, View, Text, Image, TouchableOpacity  } from 'react-native';
 import Modal from 'react-native-modal';
-import * as ImagePicker from 'expo-image-picker';
-import Constants from 'expo-constants';
-import * as Permissions from 'expo-permissions';
-import DogComponent from './DogComponent'
+import DogComponent from './DogComponent';
+import dogAvatar from '../../assets/noDogImg.jpg'
 import {getAllUserDogs} from '../../severSimulator/allUsersDog'
 import addBtn from '../../assets/PngItem_1121197.png'
 import Header from  '../../generalComponents/Header/Header'
@@ -17,18 +15,14 @@ export default class DogsScreen extends React.Component {
     this.state={
       allUserDogs:[],
       isModalVisible:false,
-      image:null
     }
     this.onAddClick=this.onAddClick.bind(this);
     this.formCallBack=this.formCallBack.bind(this);
     this.buildForm=this.buildForm.bind(this);
-    this.getPermissionAsync=this.getPermissionAsync.bind(this);
-    this.pickImage=this.pickImage.bind(this);
   }
   componentDidMount() {
     let allUserDogs=getAllUserDogs();
     this.setState({allUserDogs:allUserDogs})
-    this.getPermissionAsync();
     console.log('hi');
   }
   buildForm(){
@@ -36,24 +30,17 @@ export default class DogsScreen extends React.Component {
       {type:"text", field:"dogName", title:"dog name", labelVisibale:true},
       {type:"radio", field:"gender", title:"gender", labelVisibale:true, radioProps: [{label: 'male     ', value: 0 },{label: 'female', value: 1 }]},
       {type:'combo', field:"age", title:"age",labelVisibale:true, data:[{label:"1",value:1}, {label:"2",value:2}, {label:"3",value:3}, {label:"4",value:4}]},
-      {type:'button', field:"pic", labelVisibale:false, title:"add pic", onClick:()=>{this.pickImage()}}
+      {type:'pic', field:"dogImg", labelVisibale:false, title:"add pic"}
     ]); 
     return <Form fields={fields} callBack={this.formCallBack}></Form>
  
   }
 
-  async getPermissionAsync(){
-    if (Constants.platform.ios) {
-      const { status } = await Permissions.askAsync(Permissions.CAMERA_ROLL);
-      if (status !== 'granted') {
-        alert('Sorry, we need camera roll permissions to make this work!');
-      }
-    }
-  }
   formCallBack(fieldsToValue){
     console.log(fieldsToValue);
+    let newDogImage=fieldsToValue.dogImg?fieldsToValue.dogImg:dogAvatar;
       const newDog={
-        dogImg:{uri: this.state.image},
+        dogImg:newDogImage,
         dogName:fieldsToValue.dogName
       }
       let newAllUserDogs=[...this.state.allUserDogs];
@@ -66,23 +53,7 @@ export default class DogsScreen extends React.Component {
     const lastModalState=this.state.isModalVisible;
     this.setState({isModalVisible:!lastModalState});
   }
-  async pickImage(){
-    try {
-      let result = await ImagePicker.launchImageLibraryAsync({
-        mediaTypes: ImagePicker.MediaTypeOptions.All,
-        allowsEditing: true,
-        aspect: [4, 3],
-        quality: 1,
-      });
-      if (!result.cancelled) {
-        this.setState({ image: result.uri });
-      }
 
-      console.log(result);
-    } catch (E) {
-      console.log(E);
-    }
-  };
   render(){
     return (
       <View style={styles.container}>
